WT-17650 Fold behaviour tests into test_layered_fast_truncate19
